+++ changelog.md
@@ -1,3 +1,10 @@
+Pending Next Release
+=========
+- Update to Microsoft.Identity.Abstractions 5.2.0
+
+### New features
+- Added support for Managed Identity Federated Identity Credential. See issue [2749](https://github.com/AzureAD/microsoft-identity-web/issues/2749) for details.
+
 2.17.5
 =========
 - Updated to MSAL 4.59.1.
@@ -12,7 +19,7 @@
 2.17.3
 =========
 - Updated to Microsoft.IdentityModel.* 7.5.0
-  
+
 2.17.2
 =========